Use Yarn as package manager
Yarn provided tooling that NPM did not so many years, for example using
a lock file to download dependencies and their dependencies. Of course,
this has been changing every day, but many people already switched to
Yarn since it provided great UX for developers.

This is still a controversial topic that many people might not want to
discuss in public.